backend/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.6.0
django-postgres-extensions==0.9.3
djangorestframework==3.15.2
drf-yasg==1.21.8
et_xmlfile==2.0.0
filelock==3.16.1
greenlet==3.1.1
gunicorn==21.2.0
inflection==0.5.1
openpyxl==3.1.5
packaging==24.2
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
pywin32==308; platform_system=="Windows"
PyYAML==6.0.2
SQLAlchemy==2.0.36
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.28.0
